qcom: msm8916: Fix the no audio during device switch in voice call
During the device switch from the headphones to the speaker,
the headphones path would disable the rx chain, but while
enabling the handset path the same is not enabled back
which leads to RX interpolator to go down leading to no RX path.
Fix by powering the iir path from the handset mic as the
sidetone use case is enabled.

Change-Id: Ie9facf20b7fd20f562f0684dd87d01250f1c56bc